Improve validation for objects in deepMerge function
A new condition has been added in the deepMerge function that checks each value of the "out" object, and removes any key that has an undefined value. This was causing issues when the "out" object was used in subsequent steps, as certain as functions could not handle keys with undefined values. By deleting these keys, this issue is now mitigated.